add users page and restructure path
