+++ src/app/admin/users/page.tsx
@@ -1,0 +1,5 @@
+"use client";
+
+export default function UserListingPage() {
+  return <h1>User Listing Page</h1>;
+}
